requirements.txt
flask>=3.1.3
openai-whisper>=20250625
moviepy
numpy
pandas
tqdm
ffmpeg-python